requirements.txt
streamlit
torch
torchvision
timm
opencv-python
Pillow
requests
numpy
matplotlib
scikit-learn